Moved leveldb implementation to leveldb directory
