+++ lib/neosearch/store/store.go
@@ -1,7 +1,5 @@
 package store
 
-import "errors"
-
 // KVReader is a reader safe for concurrent reads.
 type KVReader interface {
 	Get([]byte) ([]byte, error)
@@ -54,30 +52,4 @@ type KVConfig struct {
 	DataDir     string
 	EnableCache bool
 	CacheSize   int
-}
-
-// KVFuncConstructor is the register function that every store backend need to implement.
-type KVFuncConstructor func(*KVConfig) (KVStore, error)
-
-// KVStoreConstructor is a pointer to constructor of default KVStore
-var KVStoreConstructor KVFuncConstructor
-
-// KVStoreName have the name of kv store
-var KVStoreName string
-
-// SetDefault set the default kv store
-func SetDefault(name string, initPtr KVFuncConstructor) error {
-	KVStoreName = name
-	KVStoreConstructor = initPtr
-
-	return nil
-}
-
-// New initialize the default KV store.
-func New(config *KVConfig) (KVStore, error) {
-	if KVStoreConstructor != nil {
-		return KVStoreConstructor(config)
-	}
-
-	return nil, errors.New("No store backend configured...")
 }
